Bug fixing and code improvements

+++ lib/Checkout/Common/MarketplaceCommission.php
@@ -6,6 +6,6 @@ class MarketplaceCommission
 {
     public int $amount;
 
-    public int $percentage;
+    public float $percentage;
 
 }
